scripts/requirements.txt
azure-identity
msgraph-sdk
rich
python-dotenv
